server/requirements.txt
fastapi
uvicorn
langchain
langchain-community
litellm
mcp
supabase
psycopg2-binary
python-dotenv
pydantic
httpx